// guibiometricos/src/app/module/identity/model/credential.model.ts
export class Credential {


  token?: string;
  resultado?: number;
  texto?: string;
  vigIni?: string;
  vigFin?: string;
  certificado?: string;
  rfc?: string;
  password?: string;
  serie_cert?: string;
  curp?: string;
  nombreRazonSocial?: string;


  codigo?: number;
  textmensajeo?: string;
}